Verify Cgroup Memory - cgroupfs
Refractoring to check if cgroupfs cgroup manager has enough memory
