+++ internal/config/cgmgr/cgroupfs.go
@@ -14,7 +14,10 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-type CgroupfsManager struct{}
+// CgroupfsManager defines functionality whrn **** TODO: Update this
+type CgroupfsManager struct {
+	memoryPath, memoryMaxFile string
+}
 
 const (
 	defaultCgroupfsParent = "/crio"
@@ -49,10 +52,15 @@ func (m *CgroupfsManager) ContainerCgroupAbsolutePath(sbParent, containerID stri
 
 // SandboxCgroupPath takes the sandbox parent, and sandbox ID. It
 // returns the cgroup parent, cgroup path, and error.
-func (*CgroupfsManager) SandboxCgroupPath(sbParent, sbID string) (cgParent, cgPath string, _ error) {
+func (m *CgroupfsManager) SandboxCgroupPath(sbParent, sbID string) (cgParent, cgPath string, _ error) {
 	if strings.HasSuffix(path.Base(sbParent), ".slice") {
 		return "", "", fmt.Errorf("cri-o configured with cgroupfs cgroup manager, but received systemd slice as parent: %s", sbParent)
 	}
+
+	if err := verifyCgroupHasEnoughMemory(sbParent, m.memoryPath, m.memoryMaxFile); err != nil {
+		return "", "", err
+	}
+
 	return sbParent, filepath.Join(sbParent, crioPrefix+"-"+sbID), nil
 }
